refactor: consolidate deduplication workflow by merging dedupe-split into dedupe-apply with atomic operations and validation
