+++ .gitignore
@@ -24,4 +24,3 @@ node_modules/
 dedupe-candidates.json
 dedupe-resolutions.json
 dedupe-report.md
-breweries-deduped.csv
